move field.js into 'shared'
