+++ src/admin/dumb_components/member_fields.js
@@ -1,7 +1,7 @@
 const { reduxForm } = require('redux-form')
 const { prop } = require('ramda')
 const React = require('react')
-const Field = require('../components/field.js')
+const Field = require('../../shared/dumb_components/field.js')
 const { options, field_order, fieldStructure, read_only, validate
   , is_required } =
     require('../form_fields/member.js')
