+++ src/admin/dumb_components/sub_due_form.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Field from '../components/field.js'
+import Field from '../../shared/dumb_components/field.js'
 const { reduxForm } = require('redux-form')
 
 const sub_due_dates = ({fields, handleSubmit, error}) =>
