+++ src/user/components/my_details/details_form.js
@@ -1,23 +1,36 @@
 import React from 'react'
 import { map, contains, keysIn } from 'ramda'
 import { reduxForm } from 'redux-form';
-import Field from '../../../admin/components/field.js'
+import Field from '../../../shared/dumb_components/field.js'
 
-const my_details_form = ({ ...props, read_only, edit_mode, resetForm, fields, handleSubmit, error, readOnlyFields }) => {
-  const field_creator = field =>
-    <Field
-      {...fields[field]}
-      className={!edit_mode || contains(field, readOnlyFields) ? 'user-input-read-only' : 'user-input-editable'}
-      id={field}
-      key={field}
-      mode='edit'
-      readOnly={!edit_mode || contains(field, readOnlyFields)}
-    />
+const my_details_form = (
+  { ...props
+  , toggle_edit_mode
+  , read_only
+  , edit_mode
+  , resetForm
+  , fields
+  , handleSubmit
+  , error
+  , readOnlyFields
+  }) => {
 
-  return(
+  const field_creator = field => (
+      <Field
+        {...fields[field]}
+        className={!edit_mode || contains(field, readOnlyFields) ? 'user-input-read-only' : 'user-input-editable'}
+        id={field}
+        key={field}
+        mode='edit'
+        readOnly={!edit_mode || contains(field, readOnlyFields)}
+        />
+    )
+
+
+  return (
     <form onSubmit={handleSubmit}>
       { map(field_creator)(keysIn(fields)) }
-      { edit_mode ? submit_cancel_buttons(resetForm, props.toggle_edit_mode) : edit_button(props.toggle_edit_mode) }
+      { edit_mode ? submit_cancel_buttons(resetForm, toggle_edit_mode) : edit_button(toggle_edit_mode) }
     </form>
   )
 }
